+++ Source/abci/Foundation/aiMeshOps.cpp
@@ -217,7 +217,7 @@ void MeshRefiner::retopology(bool swap_faces)
     }
 }
 
-void MeshRefiner::genSubmeshes(IArray<int> material_ids)
+void MeshRefiner::genSubmeshes(IArray<int> material_ids, std::vector<std::string> &faceset_names)
 {
     if (material_ids.empty())
     {
@@ -234,7 +234,7 @@ void MeshRefiner::genSubmeshes(IArray<int> material_ids)
 
     int num_splits = (int)splits.size();
     int offset_faces = 0;
-    RawVector<Submesh> tmp_submeshes;
+    std::vector<Submesh> tmp_submeshes;
     RawVector<int> materialOrder;
     std::unordered_set<int> materialSet;
 
@@ -289,10 +289,15 @@ void MeshRefiner::genSubmeshes(IArray<int> material_ids)
                 }
             }
 
-            for(int i=0; i < materialOrder.size(); ++i)
+            auto copyFacesetNames = materialOrder.size() == faceset_names.size(); // safety
+            for (int i = 0; i < materialOrder.size(); ++i)
             {
                 auto mi = materialOrder[i];
                 auto& sm = tmp_submeshes[mi];
+                if (copyFacesetNames)
+                {
+                    sm.facesetName = faceset_names[i];
+                }
                 if (sm.index_count > 0)
                 {
                     ++split.submesh_count;